+++ src/packages/client/store/index.ts
@@ -1270,6 +1270,7 @@ class Store
 
   addOutput(...args: Parameters<OutputActions['addOutput']>) { return this.outputActions.addOutput(...args); }
   clearOutputs(...args: Parameters<OutputActions['clearOutputs']>) { return this.outputActions.clearOutputs(...args); }
+  settleOpenStreams(...args: Parameters<OutputActions['settleOpenStreams']>) { return this.outputActions.settleOpenStreams(...args); }
   getOutputs(...args: Parameters<OutputActions['getOutputs']>) { return this.outputActions.getOutputs(...args); }
   addUserPromptToOutput(...args: Parameters<OutputActions['addUserPromptToOutput']>) { return this.outputActions.addUserPromptToOutput(...args); }
   confirmUserPromptEcho(...args: Parameters<OutputActions['confirmUserPromptEcho']>) { return this.outputActions.confirmUserPromptEcho(...args); }